Fix 1.03 - 20:52 last fix - services and Controllers works!

+++ src/main/java/tihkoff/taxi/dto/TaxiDriverEntityDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Setter
 @NoArgsConstructor
 public class TaxiDriverEntityDTO {
-    private Integer driverID;
+    private Long driverID;
 
     private String licenceNumber;
